refactor: ThemeProductGrid에서 상품 데이터 직접 쿼리하도록 구조 변경
